requirements.txt
streamlit==1.13.0
mysql-connector-python==8.0.28
ultralytics==8.0.0
opencv-python==4.5.5.64
Pillow==8.4.0
matplotlib==3.4.3
numpy==1.21.2
